+++ ydb/core/persqueue/pqtablet/partition/mlp/mlp_consumer.h
@@ -2,6 +2,7 @@
 
 #include "mlp.h"
 #include "mlp_common.h"
+#include "mlp_consumer_order.h"
 
 #include <ydb/core/base/tablet_pipecache.h>
 #include <ydb/core/keyvalue/keyvalue_events.h>
@@ -151,41 +152,8 @@ private:
     TInstant LastTimeWithMessages;
     bool LastUseForReading = false;
 
-public:
-    struct TChildPartitionsKeepOrderManager {
-    public:
-        enum class ESendReasons: ui8 {
-            None = 0,
-            Initial = 1 << 0,
-            DeliveryProblem = 1 << 1,
-            Commit = 1 << 2,
-            ParentDone = 1 << 3,
-        };
-
-        struct TChildrenPartitionWithKeepOrder {
-            ui64 TabletId = 0;
-            ui32 Cookie = 0;
-            ESendReasons SendFullStateReasons = ESendReasons::None;
-
-            bool AddSendFullStateReason(ESendReasons reason);
-            bool NeedSendFullState() const;
-        };
-
-        bool Empty() const;
-        bool SetSendFullStateToAll(ESendReasons reason);
-        bool SetSendFullStateByCookie(ui32 cookie, ESendReasons reason);
-
-        static TString SendReasonsToString(ESendReasons reason);
-
-    public:
-        ui64 ConsumerStep = 0;
-        ui32 ChildrenPartitionWithKeepOrderCookie = 2;
-
-        TMap<ui32, TChildrenPartitionWithKeepOrder> ChildrenPartitionWithKeepOrder;
-        TBackoff UpdateChildPartitionsBackoff{TDuration::Seconds(1), TDuration::Seconds(10)};
-    };
-private:
-    TChildPartitionsKeepOrderManager ChildPartitionsKeepOrderManager;
+    TChildPartitionsOrderManager ChildPartitionsOrderManager;
+    ui32 ChildrenPartitionWithKeepOrderCookie = 2;
 };
 
 class TDetailedMetrics {
